requirements.txt
pytest == 8.3.5
pip == 25.0.1
requests == 2.32.5
jsonschema == 4.25.1
allure-pytest==2.14.0
selene==2.0.0rc9
allure-python-commons~=2.14.0